Refactor common webhook logic
Move validation webhook logic to a common package, so that virt-api and
virt-operator access it both from a general package.
