+++ pkg/util/webhooks/validating-webhooks/BUILD.bazel
@@ -2,13 +2,11 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "go_default_library",
-    srcs = ["mutating-webhook.go"],
-    importpath = "kubevirt.io/kubevirt/pkg/virt-api/webhooks/mutating-webhook",
+    srcs = ["validating-webhook.go"],
+    importpath = "kubevirt.io/kubevirt/pkg/util/webhooks/validating-webhooks",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/util/webhooks:go_default_library",
-        "//pkg/virt-api/webhooks/mutating-webhook/mutators:go_default_library",
-        "//pkg/virt-config:go_default_library",
         "//staging/src/kubevirt.io/client-go/log:go_default_library",
         "//vendor/k8s.io/api/admission/v1beta1:go_default_library",
         "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
