+++ pkg/util/webhooks/webhooks_suite_test.go
@@ -1,0 +1,13 @@
+package webhooks_test
+
+import (
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	"testing"
+)
+
+func TestWebhooks(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "Webhooks Suite")
+}
