+++ pkg/virt-api/BUILD.bazel
@@ -2,10 +2,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "api.go",
-        "ca-manager.go",
-    ],
+    srcs = ["api.go"],
     importpath = "kubevirt.io/kubevirt/pkg/virt-api",
     visibility = ["//visibility:public"],
     deps = [
@@ -16,6 +13,7 @@ go_library(
         "//pkg/service:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/openapi:go_default_library",
+        "//pkg/util/webhooks:go_default_library",
         "//pkg/virt-api/rest:go_default_library",
         "//pkg/virt-api/webhooks:go_default_library",
         "//pkg/virt-api/webhooks/mutating-webhook:go_default_library",
@@ -47,7 +45,6 @@ go_test(
     name = "go_default_test",
     srcs = [
         "api_test.go",
-        "ca-manager_test.go",
         "virt-api_suite_test.go",
     ],
     embed = [":go_default_library"],
