+++ pkg/virt-api/webhooks/utils.go
@@ -20,21 +20,14 @@
 package webhooks
 
 import (
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"net/http"
 	"sync"
 
 	"github.com/golang/glog"
-	"k8s.io/api/admission/v1beta1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/tools/cache"
 
 	v1 "kubevirt.io/client-go/api/v1"
 	"kubevirt.io/client-go/kubecli"
-	"kubevirt.io/client-go/log"
 	clientutil "kubevirt.io/client-go/util"
 	"kubevirt.io/kubevirt/pkg/controller"
 	"kubevirt.io/kubevirt/pkg/util/openapi"
@@ -74,19 +67,6 @@ var MigrationGroupVersionResource = metav1.GroupVersionResource{
 	Group:    v1.VirtualMachineInstanceMigrationGroupVersionKind.Group,
 	Version:  v1.VirtualMachineInstanceMigrationGroupVersionKind.Version,
 	Resource: "virtualmachineinstancemigrations",
-}
-
-func ValidateRequestResource(request metav1.GroupVersionResource, group string, resource string) bool {
-	gvr := metav1.GroupVersionResource{Group: group, Resource: resource}
-
-	for _, version := range v1.ApiSupportedWebhookVersions {
-		gvr.Version = version
-		if gvr == request {
-			return true
-		}
-	}
-
-	return false
 }
 
 type Informers struct {
@@ -129,85 +109,4 @@ func newInformers() *Informers {
 		VMIPresetInformer:       kubeInformerFactory.VirtualMachinePreset(),
 		NamespaceLimitsInformer: kubeInformerFactory.LimitRanges(),
 	}
-}
-
-// GetAdmissionReview
-func GetAdmissionReview(r *http.Request) (*v1beta1.AdmissionReview, error) {
-	var body []byte
-	if r.Body != nil {
-		if data, err := ioutil.ReadAll(r.Body); err == nil {
-			body = data
-		}
-	}
-
-	// verify the content type is accurate
-	contentType := r.Header.Get("Content-Type")
-	if contentType != "application/json" {
-		return nil, fmt.Errorf("contentType=%s, expect application/json", contentType)
-	}
-
-	ar := &v1beta1.AdmissionReview{}
-	err := json.Unmarshal(body, ar)
-	return ar, err
-}
-
-// ToAdmissionResponseError
-func ToAdmissionResponseError(err error) *v1beta1.AdmissionResponse {
-	log.Log.Reason(err).Error("admission generic error")
-
-	return &v1beta1.AdmissionResponse{
-		Result: &metav1.Status{
-			Message: err.Error(),
-			Code:    http.StatusBadRequest,
-		},
-	}
-}
-
-func ToAdmissionResponse(causes []metav1.StatusCause) *v1beta1.AdmissionResponse {
-	log.Log.Infof("rejected vmi admission")
-
-	globalMessage := ""
-	for _, cause := range causes {
-		if globalMessage == "" {
-			globalMessage = cause.Message
-		} else {
-			globalMessage = fmt.Sprintf("%s, %s", globalMessage, cause.Message)
-		}
-	}
-
-	return &v1beta1.AdmissionResponse{
-		Result: &metav1.Status{
-			Message: globalMessage,
-			Reason:  metav1.StatusReasonInvalid,
-			Code:    http.StatusUnprocessableEntity,
-			Details: &metav1.StatusDetails{
-				Causes: causes,
-			},
-		},
-	}
-}
-
-func ValidationErrorsToAdmissionResponse(errs []error) *v1beta1.AdmissionResponse {
-	var causes []metav1.StatusCause
-	for _, e := range errs {
-		causes = append(causes,
-			metav1.StatusCause{
-				Message: e.Error(),
-			},
-		)
-	}
-	return ToAdmissionResponse(causes)
-}
-
-func ValidateSchema(gvk schema.GroupVersionKind, data []byte) *v1beta1.AdmissionResponse {
-	in := map[string]interface{}{}
-	err := json.Unmarshal(data, &in)
-	if err != nil {
-		return ToAdmissionResponseError(err)
-	}
-	errs := Validator.Validate(gvk, in)
-	if len(errs) > 0 {
-		return ValidationErrorsToAdmissionResponse(errs)
-	}
-	return nil
 }
