+++ pkg/virt-api/webhooks/validating-webhook/BUILD.bazel
@@ -6,12 +6,9 @@ go_library(
     importpath = "kubevirt.io/kubevirt/pkg/virt-api/webhooks/validating-webhook",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/virt-api/webhooks:go_default_library",
+        "//pkg/util/webhooks/validating-webhooks:go_default_library",
         "//pkg/virt-api/webhooks/validating-webhook/admitters:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//staging/src/kubevirt.io/client-go/kubecli:go_default_library",
-        "//staging/src/kubevirt.io/client-go/log:go_default_library",
-        "//vendor/k8s.io/api/admission/v1beta1:go_default_library",
-        "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
     ],
 )
