+++ pkg/virt-api/webhooks/validating-webhook/admitters/BUILD.bazel
@@ -17,6 +17,7 @@ go_library(
         "//pkg/hooks:go_default_library",
         "//pkg/util:go_default_library",
         "//pkg/util/hardware:go_default_library",
+        "//pkg/util/webhooks:go_default_library",
         "//pkg/virt-api/webhooks:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-operator/creation/rbac:go_default_library",
